Move helpers from tslib.es6.js to modules/index.js
